Add ignore option to kplutus-cover
* The ignore option allows the desciption of files containing rules
  that will be disregarded by in the coverage analysis. This is useful
  for instance for files containing rule macros or functions that are
  not relevant to a rule coverage analysis.

* `Makefile` was updated accordingly.

+++ Makefile
@@ -266,8 +266,22 @@ $(KPLUTUS_LIB)/$(haskell_kompiled): $(kplutus_includes) $(plugin_includes) $(KPL
 # Coverage Processing
 # -------------------
 
+covr_ignore_files := bitstream.md \
+                     uplc-bytestring.md \
+                     uplc-flat-parser.md
+
+covr_ignore_includes := $(patsubst %, $(KPLUTUS_INCLUDE)/kframework/%, $(covr_ignore_files))
+
+K_BUILTINS_INCLUDE := $(KPLUTUS_LIB)/kframework/include/kframework/builtin
+
+k_builtin_ignore_files := domains.md
+
+k_builtin_ignore_includes := $(patsubst %, $(K_BUILTINS_INCLUDE)/%, $(k_builtin_ignore_files))
+
 coverage:
-	$(KPLUTUS_BIN)/kplutus-covr $(KPLUTUS_LIB)/$(llvm_kompiled_dir) -- $(kplutus_includes) -nd > $(BUILD_DIR)/coverage.xml
+	$(KPLUTUS_BIN)/kplutus-covr $(KPLUTUS_LIB)/$(llvm_kompiled_dir) \
+        -- $(kplutus_includes) \
+        -ig $(covr_ignore_includes) $(k_builtin_ignore_includes) > $(BUILD_DIR)/coverage.xml
 
 # Installing
 # ----------
@@ -335,7 +349,7 @@ uninstall:
 procs := $(shell nproc)
 
 fresh-test-coverage:
-	rm -r $(KPLUTUS_LIB)/$(llvm_kompiled_dir)
+	[ -d $(KPLUTUS_LIB)/$(llvm_kompiled_dir) ] && rm -r $(KPLUTUS_LIB)/$(llvm_kompiled_dir)
 	make build-coverage
 	make conformance-test -j$(procs)
 	make coverage
